Change structure of Tile
The player that owns the stone on a tile is now referenced directly and
no longer by number. The neighbours and tileProp fields are also renamed
to transitions respectively property, both being more clearly.

A Property enum is (re)introduced.

JavaDoc comments are added to all methods and important fields.

The Direction enum is added, instances of which represent on of eight
directions.