Fix BuildMove doMove for non-straight lines
The direction in which the transition arrives
determines the direction of search.

ArrayList is now initialized in the for loop.

Steps is replaced with direct reference to last
tile in path. This saves a few gets. Step count
can still be obtained with size().

searchDirection for array indexing is now directly
done via directions.ordinal().