Move MoveType to Move class
MoveType is better placed in the Move class itself,
because it can be reused from there and doesn't have
to be redeclared if it is needed elsewhere in the
code.

Also we don't need typeof anymore to determine the
type of a move.

Every occurrence of MoveType can be replaced with
Move.Type, along with the import of bacon.move.Move.