Complete Overhaul
The previous implementation of aspiration window was incorrect in working principle. Instead initializing alpha and beta with aspiration window boundaries, it is the initial value of every node that needs to be initialized with aspiration window boundaries.

Furthermore, the variable boolean windowSuccess is introduced to help identify and discard empty nodes. This version now both shows correct behavior AND plays well.